consumer-agent/requirements.txt
stellar-sdk>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.40.0
